Rename saveMetricsInQueryParameters to updateQueryParameters #2498
