+++ visualization/app/codeCharta/services/loadInitialFile/loadInitialFile.service.spec.ts
@@ -8,7 +8,7 @@ import stringify from "safe-stable-stringify"
 import { AppSettings, CcState, DynamicSettings, FileSettings } from "../../codeCharta.model"
 import { FileSelectionState } from "../../model/files/files"
 import { getCCFiles } from "../../model/files/files.helper"
-import { MetricQueryParemter } from "../../state/effects/saveMetricsInQueryParameters/saveMetricsInQueryParameters.effect"
+import { MetricQueryParemter } from "../../state/effects/updateQueryParameters/updateQueryParameters.effect"
 import { metricDataSelector } from "../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { setAmountOfTopLabels } from "../../state/store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
 import { defaultAppSettings } from "../../state/store/appSettings/appSettings.reducer"
