+++ visualization/app/codeCharta/services/loadInitialFile/loadInitialFile.service.ts
@@ -47,7 +47,7 @@ import { ExportCCFile } from "../../codeCharta.api.model"
 import { AppSettings, CcState, DynamicSettings, FileSettings, NameDataPair } from "../../codeCharta.model"
 import { FileState } from "../../model/files/files"
 import { getCCFiles } from "../../model/files/files.helper"
-import { MetricQueryParemter } from "../../state/effects/saveMetricsInQueryParameters/saveMetricsInQueryParameters.effect"
+import { MetricQueryParemter } from "../../state/effects/updateQueryParameters/updateQueryParameters.effect"
 import { metricDataSelector } from "../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { setIsLoadingFile } from "../../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
 import { setIsLoadingMap } from "../../state/store/appSettings/isLoadingMap/isLoadingMap.actions"
